chore: move foreign call integration tests into proper integration test files
